ci: Use `actions/setup-node` to remove a dependency (#643)
This repository currently uses both `guardian/actions-setup-node` and `actions/setup-node`.

`guardian/actions-setup-node` is a fork of `actions/setup-node` and has fallen behind. In this change, we move to using only `actions/setup-node` for simplicity.

Also replace `bahmutov/npm-install` with a DIY solution, as it's not too complex and removes a dependency: we use `actions/setup-node` to manage the cache, and manually install yarn deps.